requirements.txt
flask==2.0.1
beautifulsoup4==4.10.0
spacy==3.2.0
pymongo==4.0.1
python-jwt==3.3.0
python-dotenv==0.19.2
gunicorn==20.1.0